Extract inventory safety policy
Move repository inventory detection and safe-pipeline checks out of command parsing. Import the policy directly from its focused module and keep the existing coverage in a matching test file.
