+++ src/core/command.ts
@@ -3,103 +3,10 @@ import { basename } from "node:path";
 import type { ToolExecutionInput } from "../types.js";
 
 type ShellOperator = ";" | "\n" | "|" | "&&" | "||";
-type StdinFilterSpec = {
-  flags?: readonly string[];
-  valueFlags?: readonly string[];
-  inlineValuePrefixes?: readonly string[];
-  compactValueFlags?: readonly RegExp[];
-};
 
 const FILE_CONTENT_INSPECTION_COMMANDS = new Set(["cat", "sed", "head", "tail", "nl", "bat", "batcat", "jq", "yq"]);
-const REPO_INVENTORY_COMMANDS = new Set(["find", "fd", "fdfind", "ls"]);
-const FIND_EXEC_ACTIONS = new Set(["-exec", "-execdir", "-ok", "-okdir"]);
-const FD_EXEC_OPTIONS = new Set(["-x", "--exec", "-X", "--exec-batch"]);
 const COMPOUND_SHELL_OPERATORS = new Set<ShellOperator>([";", "\n", "|", "&&", "||"]);
 const SEQUENTIAL_SHELL_OPERATORS = new Set<ShellOperator>([";", "\n", "&&", "||"]);
-const HEAD_FILTER: StdinFilterSpec = {
-  flags: ["-q", "-v", "-z"],
-  valueFlags: ["-n", "-c", "--lines", "--bytes"],
-  inlineValuePrefixes: ["--lines=", "--bytes="],
-  compactValueFlags: [/^-[0-9]+$/u, /^-[nc].+/u],
-};
-const TAIL_FILTER: StdinFilterSpec = {
-  flags: ["-q", "-v", "-z", "-f", "-F", "-r"],
-  valueFlags: ["-n", "-c", "--lines", "--bytes", "--pid", "-s", "--sleep-interval"],
-  inlineValuePrefixes: ["--lines=", "--bytes=", "--pid=", "--sleep-interval="],
-  compactValueFlags: [/^-[0-9]+$/u, /^-[nc].+/u],
-};
-const SORT_FILTER: StdinFilterSpec = {
-  flags: [
-    "-b",
-    "-d",
-    "-f",
-    "-g",
-    "-h",
-    "-i",
-    "-M",
-    "-m",
-    "-n",
-    "-R",
-    "-r",
-    "-s",
-    "-u",
-    "-V",
-    "-z",
-    "--check",
-    "--debug",
-    "--dictionary-order",
-    "--general-numeric-sort",
-    "--human-numeric-sort",
-    "--ignore-case",
-    "--ignore-leading-blanks",
-    "--ignore-nonprinting",
-    "--merge",
-    "--month-sort",
-    "--numeric-sort",
-    "--random-sort",
-    "--reverse",
-    "--stable",
-    "--unique",
-    "--version-sort",
-    "--zero-terminated",
-  ],
-  valueFlags: ["-k", "-S", "-t", "--batch-size", "--key", "--sort"],
-  inlineValuePrefixes: ["--batch-size=", "--key=", "--sort="],
-  compactValueFlags: [/^-[bdfghinMmrRsuVz]*k.+/u, /^-[bdfghinMmrRsuVz]*S.+/u, /^-[bdfghinMmrRsuVz]*t.+/u],
-};
-const UNIQ_FILTER: StdinFilterSpec = {
-  flags: [
-    "-c",
-    "-d",
-    "-D",
-    "-i",
-    "-u",
-    "-z",
-    "--all-repeated",
-    "--count",
-    "--ignore-case",
-    "--repeated",
-    "--unique",
-    "--zero-terminated",
-  ],
-  valueFlags: ["-f", "-s", "-w", "--skip-fields", "--skip-chars", "--check-chars"],
-  inlineValuePrefixes: ["--skip-fields=", "--skip-chars=", "--check-chars=", "--all-repeated="],
-  compactValueFlags: [/^-[fsw].+/u],
-};
-const REPO_INVENTORY_PIPE_FILTERS = new Map<string, StdinFilterSpec>([
-  ["head", HEAD_FILTER],
-  ["sort", SORT_FILTER],
-  ["tail", TAIL_FILTER],
-  ["uniq", UNIQ_FILTER],
-]);
-
-export type RepositoryInventorySafety = "not-inventory" | "safe" | "sequential-command" | "unsafe-pipeline";
-export type InspectionCommandPolicy = "compact-all" | "skip-all" | "skip-file-content" | "allow-safe-inventory";
-export type InspectionCommandSkipReason =
-  | "inspection-command"
-  | "file-content-inspection-command"
-  | "sequential-inventory-command"
-  | "unsafe-inventory-pipeline";
 
 function getNormalizedArgv(input: Pick<ToolExecutionInput, "argv" | "command">): string[] {
   if (input.argv?.length) {
@@ -111,7 +18,7 @@ function getNormalizedArgv(input: Pick<ToolExecutionInput, "argv" | "command">):
   return tokenizeCommand(stripLeadingCdPrefix(input.command));
 }
 
-function getNormalizedArgv0(argv: string[]): string | null {
+export function getCommandName(argv: string[]): string | null {
   const first = argv[0];
   if (!first) {
     return null;
@@ -218,108 +125,6 @@ function hasUnquotedShellOperator(command: string, operators: Set<ShellOperator>
   return false;
 }
 
-function splitUnquotedPipes(command: string): string[] {
-  const segments: string[] = [];
-  let current = "";
-  let quote: "'" | "\"" | null = null;
-  let escaping = false;
-
-  for (let index = 0; index < command.length; index += 1) {
-    const char = command[index]!;
-
-    if (escaping) {
-      current += char;
-      escaping = false;
-      continue;
-    }
-
-    if (char === "\\") {
-      current += char;
-      escaping = true;
-      continue;
-    }
-
-    if (quote) {
-      current += char;
-      if (char === quote) {
-        quote = null;
-      }
-      continue;
-    }
-
-    if (char === "'" || char === "\"") {
-      current += char;
-      quote = char;
-      continue;
-    }
-
-    if (char === "|" && command[index + 1] !== "|") {
-      if (current.trim()) {
-        segments.push(current.trim());
-      }
-      current = "";
-      continue;
-    }
-
-    current += char;
-  }
-
-  if (current.trim()) {
-    segments.push(current.trim());
-  }
-
-  return segments;
-}
-
-function isKnownFlag(arg: string, spec: StdinFilterSpec): boolean {
-  return (spec.flags?.includes(arg) ?? false)
-    || (spec.inlineValuePrefixes?.some((prefix) => arg.startsWith(prefix)) ?? false)
-    || (spec.compactValueFlags?.some((pattern) => pattern.test(arg)) ?? false);
-}
-
-function isStdinOnlyFilter(argv: string[], spec: StdinFilterSpec): boolean {
-  for (let index = 1; index < argv.length; index += 1) {
-    const arg = argv[index]!;
-    if (arg === "-") {
-      continue;
-    }
-    if (arg === "--") {
-      return argv.slice(index + 1).every((operand) => operand === "-");
-    }
-    if (spec.valueFlags?.includes(arg)) {
-      index += 1;
-      if (index >= argv.length) {
-        return false;
-      }
-      continue;
-    }
-    if (isKnownFlag(arg, spec)) {
-      continue;
-    }
-    return false;
-  }
-  return true;
-}
-
-function isSafeRepositoryInventoryPipeSegment(commandName: string, argv: string[]): boolean {
-  const spec = REPO_INVENTORY_PIPE_FILTERS.get(commandName);
-  if (!spec) {
-    return false;
-  }
-  return isStdinOnlyFilter(argv, spec);
-}
-
-function isSafeRepositoryInventorySource(argv: string[]): boolean {
-  const commandName = getNormalizedArgv0(argv);
-  if (commandName === "find") {
-    return argv.every((arg) => !FIND_EXEC_ACTIONS.has(arg));
-  }
-  if (commandName === "fd" || commandName === "fdfind") {
-    return argv.every((arg) => !FD_EXEC_OPTIONS.has(arg) && !arg.startsWith("--exec=") && !arg.startsWith("--exec-batch="));
-  }
-  return true;
-}
-
 export function isCompoundShellCommand(command: string): boolean {
   return hasUnquotedShellOperator(command, COMPOUND_SHELL_OPERATORS);
 }
@@ -349,7 +154,7 @@ function isGitGlobalOptionWithInlineValue(option: string): boolean {
 }
 
 export function getGitSubcommand(argv: string[]): string | null {
-  if (getNormalizedArgv0(argv) !== "git") {
+  if (getCommandName(argv) !== "git") {
     return null;
   }
 
@@ -380,99 +185,11 @@ export function getGitSubcommand(argv: string[]): string | null {
 
 export function isFileContentInspectionCommand(input: Pick<ToolExecutionInput, "argv" | "command">): boolean {
   const argv = getNormalizedArgv(input);
-  const argv0 = getNormalizedArgv0(argv);
+  const argv0 = getCommandName(argv);
   if (!argv0) {
     return false;
   }
   return FILE_CONTENT_INSPECTION_COMMANDS.has(argv0);
-}
-
-export function isRepositoryInventoryCommand(input: Pick<ToolExecutionInput, "argv" | "command">): boolean {
-  const argv = getNormalizedArgv(input);
-  const argv0 = getNormalizedArgv0(argv);
-  if (!argv0) {
-    return false;
-  }
-  if (REPO_INVENTORY_COMMANDS.has(argv0)) {
-    return true;
-  }
-  if (argv0 === "rg" && argv.includes("--files")) {
-    return true;
-  }
-  if (argv0 === "git" && getGitSubcommand(argv) === "ls-files") {
-    return true;
-  }
-  return false;
-}
-
-export function getRepositoryInventorySafety(command: string): RepositoryInventorySafety {
-  const effectiveCommand = stripLeadingCdPrefix(command);
-  const sourceArgv = tokenizeCommand(effectiveCommand);
-  if (!isRepositoryInventoryCommand({ argv: sourceArgv })) {
-    return "not-inventory";
-  }
-  if (!isSafeRepositoryInventorySource(sourceArgv)) {
-    return "unsafe-pipeline";
-  }
-  if (hasSequentialShellCommands(effectiveCommand)) {
-    return "sequential-command";
-  }
-
-  const segments = splitUnquotedPipes(effectiveCommand);
-  if (segments.length <= 1) {
-    return "safe";
-  }
-
-  const safePipeline = segments.slice(1).every((segment) => {
-    const argv = tokenizeCommand(segment);
-    const commandName = getNormalizedArgv0(argv);
-    return commandName !== null && isSafeRepositoryInventoryPipeSegment(commandName, argv);
-  });
-  return safePipeline ? "safe" : "unsafe-pipeline";
-}
-
-export function isSafeRepositoryInventoryPipeline(command: string): boolean {
-  return getRepositoryInventorySafety(command) === "safe";
-}
-
-export function isRepositoryInspectionCommand(input: Pick<ToolExecutionInput, "argv" | "command">): boolean {
-  if (isFileContentInspectionCommand(input)) {
-    return true;
-  }
-  if (isRepositoryInventoryCommand(input)) {
-    return true;
-  }
-  return false;
-}
-
-export function getInspectionCommandSkipReason(
-  command: string,
-  policy: InspectionCommandPolicy,
-): InspectionCommandSkipReason | null {
-  if (policy === "compact-all") {
-    return null;
-  }
-
-  if (policy === "skip-all") {
-    return isRepositoryInspectionCommand({ command }) ? "inspection-command" : null;
-  }
-
-  if (isFileContentInspectionCommand({ command })) {
-    return "file-content-inspection-command";
-  }
-
-  if (policy === "skip-file-content") {
-    return null;
-  }
-
-  const inventorySafety = getRepositoryInventorySafety(command);
-  if (inventorySafety === "sequential-command") {
-    return "sequential-inventory-command";
-  }
-  if (inventorySafety === "unsafe-pipeline") {
-    return "unsafe-inventory-pipeline";
-  }
-  return null;
 }
 
 export function normalizeCommandSignature(command?: string): string | null {
@@ -485,7 +202,7 @@ export function normalizeCommandSignature(command?: string): string | null {
     return null;
   }
 
-  const normalized = getNormalizedArgv0(argv);
+  const normalized = getCommandName(argv);
   return normalized || null;
 }
